Add files via upload
Added footer and header to home page. Made visual changes to topnavbar. Created an animation for topnavbar where if the browser detects scrolling it switches it with a dropdown menu button for navigation with a very basic animation. Fixed a visual bug with the card animation flipping where the text would sometimes not flip back to its original state.